requirements.txt
sphinx-autodoc-typehints==1.11.1
sphinx-rtd-theme==0.5.0
autodoc==0.5.0
Sphinx==3.3.1
attrs==20.3.0
numpy==1.19.4
Pillow==8.0.1
matplotlib==3.3.3
Faker==6.4.1
factory-boy==3.2.0
mockito==1.2.2
pytest-mockito==0.0.4
pytest-mock==3.6.1
